# docs/openapi.yml
# THIS FILE IS GENERATED BY OpenApiSchemaExportTest — DO NOT EDIT MANUALLY.
# Run `./mvnw -pl service -am test -Dtest=OpenApiSchemaExportTest -Dsurefire.failIfNoSpecifiedTests=false` to regenerate.
---
openapi: 3.1.0
components:
  schemas:
    ApplicationSpec:
      type: object
      deprecated: true
      required:
      - repoId
      - name
      properties:
        repoId:
          type: string
          pattern: \S
        name:
          type: string
          pattern: \S
        healthPath:
          type: string
    CdApplicationDto:
      type: object
      properties:
        id:
          type: string
        repoId:
          type: string
        name:
          type: string
        environmentId:
          type: string
        environmentName:
          type: string
        target:
          $ref: "#/components/schemas/CdDeploymentTarget"
        availableOnEnv:
          type: boolean
        branch:
          type: string
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
    CdDeploymentDto:
      type: object
      properties:
        id:
          type: string
        applicationId:
          type: string
        applicationName:
          type: string
        commitSha:
          type: string
        runId:
          type: string
        status:
          $ref: "#/components/schemas/CdDeploymentStatus"
        containerName:
          type: string
        detail:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        finishedAt:
          $ref: "#/components/schemas/Instant"
    CdDeploymentStatus:
      type: string
      enum:
      - QUEUED
      - STARTING
      - ACTIVE
      - IMAGE_MISSING
      - FAILED
      - DECOMMISSIONED
    CdDeploymentTarget:
      type: string
      enum:
      - ENVIRONMENT
      - SINGLETON
    CdEnvironmentDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        branch:
          type: string
        network:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
        applications:
          type: array
          items:
            $ref: "#/components/schemas/CdApplicationDto"
    CdPinDto:
      type: object
      properties:
        applicationName:
          type: string
        shas:
          type: array
          items:
            type: string
    CreateEnvironmentRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
        branch:
          type: string
        network:
          type: string
        applications:
          type: array
          deprecated: true
          items:
            $ref: "#/components/schemas/ApplicationSpec"
    EnvironmentResponse:
      type: object
      properties:
        environment:
          $ref: "#/components/schemas/CdEnvironmentDto"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ListApplicationsResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/CdApplicationDto"
    ListDeploymentsResponse:
      type: object
      properties:
        deployments:
          type: array
          items:
            $ref: "#/components/schemas/CdDeploymentDto"
    ListEnvironmentsResponse:
      type: object
      properties:
        environments:
          type: array
          items:
            $ref: "#/components/schemas/CdEnvironmentDto"
    ListPinsResponse:
      type: object
      properties:
        pins:
          type: array
          items:
            $ref: "#/components/schemas/CdPinDto"
    UpdateEnvironmentRequest:
      type: object
      properties:
        name:
          type: string
        branch:
          type: string
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
  /cd/api/applications:
    get:
      summary: Every application cd deploys — environment applications and singletons
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListApplicationsResponse"
      tags:
      - Cd Application Controller
  /cd/api/deployments:
    get:
      summary: "An environment's recorded deployments, newest-first"
      parameters:
      - name: environmentId
        in: query
        schema:
          type: string
      responses:
        "200":
          description: The deployments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDeploymentsResponse"
        "404":
          description: No such environment
      tags:
      - Cd Deployment Controller
  /cd/api/environments:
    get:
      summary: List environments
      responses:
        "200":
          description: The environments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvironmentsResponse"
        "502":
          description: qits-serviceregistry could not be reached
      tags:
      - Cd Environment Controller
    post:
      summary: "Create an environment: a name, a branch to listen to, a docker network"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentRequest"
        required: true
      responses:
        "201":
          description: Created; green builds on the branch now deploy
        "400":
          description: "A name, branch or path failed validation"
        "409":
          description: An environment of that name already exists
        "502":
          description: qits-serviceregistry could not be reached
      tags:
      - Cd Environment Controller
  /cd/api/environments/{environmentId}:
    patch:
      summary: Rename an environment or point it at another branch
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvironmentRequest"
        required: true
      responses:
        "200":
          description: The updated environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "400":
          description: A name or branch failed validation
        "404":
          description: No such environment
        "409":
          description: Another environment already has that name
        "502":
          description: qits-serviceregistry could not be reached
      tags:
      - Cd Environment Controller
    get:
      summary: One environment with the applications it tracks
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "404":
          description: No such environment
        "502":
          description: qits-serviceregistry could not be reached
      tags:
      - Cd Environment Controller
    delete:
      summary: "Tear an environment down (rows, containers, network)"
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Torn down
        "404":
          description: No such environment
        "502":
          description: qits-serviceregistry could not be reached
      tags:
      - Cd Environment Controller
  /cd/api/pins:
    get:
      summary: "The image shas deployments pin: what serves, and what a rollback restores"
      responses:
        "200":
          description: One entry per application name serving somewhere; an application
            with no ACTIVE deployment pins nothing and is absent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPinsResponse"
      tags:
      - Cd Pin Controller
info:
  title: qits-cd API
  version: 2026.802.191319